+++ src/routes/(editor)/modules/widgets/section/sectionoptions.svelte
@@ -1,1 +1,0 @@
-<div>Section Options</div>