+++ src/routes/(modules)/+layout@.svelte
@@ -1,0 +1,5 @@
+<script>
+    
+</script>
+<div>Modules</div>
+<slot></slot>